% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividir.R
\name{dividir}
\alias{dividir}
\title{dividir return the ratio between x and y}
\usage{
dividir(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
A numeric value from x / y
}
\description{
dividir return the ratio between x and y
}
\examples{
dividir(1, 1)

}
