% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicar.R
\name{multiplicar}
\alias{multiplicar}
\title{subtrair return the product between x and y}
\usage{
multiplicar(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
A numeric value from x * y
}
\description{
subtrair return the product between x and y
}
\examples{
multiplicar(1, 1)

}
