% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{somar return the sum of x and y}
\usage{
somar(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
A numeric value from x + y
}
\description{
somar return the sum of x and y
}
\examples{
somar(1, 1)

}
