% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{subtrair return the difference between x and y}
\usage{
subtrair(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
A numeric value from x - y
}
\description{
subtrair return the difference between x and y
}
\examples{
subtrair(1, 1)

}
